Intro to Linear Interpolation
In this program, I learned how to use linear interpolation to be able to find the value of y on a line if the user inputs an x value. The program also has a part of having to find the distance between the International Space Station from Houston using the circumference of the earth.